Redesign player class and separate functions.
