+++ src/kr/ac/snu/sbkim28/game/io/ConsoleInputPlayer.java
@@ -1,4 +1,8 @@
-package kr.ac.snu.sbkim28.game.core;
+package kr.ac.snu.sbkim28.game.io;
+
+import kr.ac.snu.sbkim28.game.core.GameOverState;
+import kr.ac.snu.sbkim28.game.core.Player;
+import kr.ac.snu.sbkim28.game.core.WordResultState;
 
 import java.io.BufferedReader;
 import java.io.IOException;
@@ -11,7 +15,7 @@ import java.io.InputStreamReader;
  * @version 1.0
  * @see Player
  */
-public class ConsoleInputPlayer extends AbstractPlayer{
+public class ConsoleInputPlayer extends AbstractDirectPlayer {
 
     /**
      * 콘솔 입력을 받는 Reader 클래스.
@@ -40,10 +44,10 @@ public class ConsoleInputPlayer extends AbstractPlayer{
     /**
      * 사용자로부터 콘솔을 통해서 입력을 받고 이를 반환함.
      * 메써드가 호출되었을 때 사용자로부터 입력을 받기까지 대기하게 됨.
-     * {@link Player#notifyGameOver()} method가 호출되면
+     * {@link Player#notifyGameOver(GameOverState)} method가 호출되면
      * 입력 대기를 중단하고 null을 반환함.
      * @return 사용자로부터 입력받은 word.
-     * 사용자 입력 대기를 받는 도중 {@link Player#notifyGameOver()}가 호출되면
+     * 사용자 입력 대기를 받는 도중 {@link Player#notifyGameOver(GameOverState)}가 호출되면
      * 대기를 멈추고 null을 반환함.
      */
     @Override
@@ -71,5 +75,15 @@ public class ConsoleInputPlayer extends AbstractPlayer{
     @Override
     public void notifyGameOver(GameOverState state) {
         waitForUserInput = false;
+    }
+
+    @Override
+    protected void createInputReceiver() {
+
+    }
+
+    @Override
+    protected void createDisplayer() {
+
     }
 }
