+++ src/kr/ac/snu/sbkim28/game/io/GameDisplayer.java
@@ -1,0 +1,4 @@
+package kr.ac.snu.sbkim28.game.io;
+
+public interface GameDisplayer {
+}
