+++ src/kr/ac/snu/sbkim28/game/io/PlayerInputReceiver.java
@@ -1,0 +1,6 @@
+package kr.ac.snu.sbkim28.game.io;
+
+public interface PlayerInputReceiver {
+    String waitUntilPlayerInput();
+    void haltWaiting();
+}
